File system tricks to make it easier to find release notes by search
