+++ front/app/containers/ReleaseNotes/version.tsx
@@ -1,0 +1,37 @@
+import * as React from 'react';
+import styled from 'styled-components';
+import Heading from 'components/Heading';
+import { Row, Col } from 'react-bootstrap';
+import RichTextEditor, { EditorValue } from 'react-rte-yt';
+
+const notes = 
+  RichTextEditor.createValueFromString(`
+
+# ClinWiki Version History
+
+### Version 2
+- Add /voyager to debug builds for exploring graphql schema
+
+### Version 1
+- Filter facet values in site config
+- Beta added to logo
+- Dynamic study sections
+- Country facet
+- Release notes added
+- Show record count on study page
+- First/Last buttons on study page
+
+### Version 0
+- Everything else
+`, 'markdown');
+
+export default class ReleaseNotes extends React.PureComponent<{},{}> {
+    render() {
+        return (
+        <div>
+            <RichTextEditor
+                readOnly
+                value={notes} />
+        </div>);
+    }
+}
